feat: testes de repository, refactor PedidosScreenViewModel, WelcomeScreenViewModel e correções de tipos
- Adicionado ComprasRepositoryTest (6 testes: CRUD + somarComprasPorPeriodo)
- Adicionado WelcomeScreenViewModel (WelcomeScreen movido para pacote welcomeScreen/)
- Refatorado PedidosScreenViewModel: remove dependência direta de PedidoItemRepository, usa PedidoItemService
- Adicionado PedidoItemService.listarPorPedido()
- Migration V7: corrigido dataCriacao INTEGER→REAL, data_compra/desconto_em_reais/preco_compra TEXT→REAL
- CompraModel.fornecedorId: Long → Integer (consistente com FornecedorModel)
- CompraDto.fornecedorId: Long → Integer
- ContasPagarService: simplificado setFornecedorId

+++ src/main/java/my_app/db/dto/CompraDto.java
@@ -5,7 +5,7 @@ import java.math.BigDecimal;
 public record CompraDto(
         String produtoCod,
         BigDecimal precoCompra,
-        Long fornecedorId,
+        Integer fornecedorId,
         BigDecimal quantidade,
         BigDecimal descontoEmReais,
         String tipoPagamento,
